src/main/java/wooteco/subway/dao/SectionDao.java
package wooteco.subway.dao;

import javax.sql.DataSource;
import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
import org.springframework.jdbc.core.namedparam.SqlParameterSource;
import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
import wooteco.subway.domain.Line;
import wooteco.subway.domain.Section;

public class SectionDao {

    private final NamedParameterJdbcTemplate jdbcTemplate;
    private final SimpleJdbcInsert simpleInsert;

    public SectionDao(DataSource dataSource) {
        this.jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
        this.simpleInsert = new SimpleJdbcInsert(dataSource)
                .withTableName("SECTION")
                .usingGeneratedKeyColumns("id");
    }

    public Section save(Line line, Section section) {
        SqlParameterSource parameters = new MapSqlParameterSource("line_id", line.getId())
                .addValue("up_station_id", section.getUpStation().getId())
                .addValue("down_station_id", section.getDownStation().getId())
                .addValue("distance", section.getDistance());
        Long id = simpleInsert.executeAndReturnKey(parameters).longValue();
        return new Section(id, section.getUpStation(), section.getDownStation(), section.getDistance());
    }

}
